requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0
redis>=5.0.0
openpyxl>=3.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
jinja2>=3.1.0
requests>=2.31.0 